Extract shared Interactivity API store for actor list blocks.
Create a factory function that generates the store configuration,
allowing followers and following blocks to share the same logic
while maintaining separate store names.
