+++ src/shared/actor-list/store.js
@@ -1,0 +1,157 @@
+import { store, getContext, getConfig } from '@wordpress/interactivity';
+
+/**
+ * @member {Object} window.wp WordPress global object
+ * @member {Function} url.addQueryArgs Function to add query arguments to a URL.
+ */
+const { apiFetch, url } = window.wp;
+
+/**
+ * Validates a URL to ensure it uses a safe scheme (http/https).
+ *
+ * @param {string} urlString The URL to validate.
+ * @return {string} The validated URL or empty string if invalid.
+ */
+function validateUrl( urlString ) {
+	try {
+		const parsed = new URL( urlString );
+		return [ 'http:', 'https:' ].includes( parsed.protocol ) ? urlString : '';
+	} catch {
+		return '';
+	}
+}
+
+/**
+ * Creates and registers an Interactivity API store for actor lists.
+ *
+ * @param {string} storeName The name of the store (e.g., 'activitypub/followers').
+ * @return {Object} The store actions object.
+ */
+export function createActorListStore( storeName ) {
+	const { actions } = store( storeName, {
+		/**
+		 * @typedef {Object} state
+		 * @property {Function} paginationText      Get the pagination text.
+		 * @property {Function} disablePreviousLink Whether the previous link should be disabled.
+		 * @property {Function} disableNextLink     Whether the next link should be disabled.
+		 */
+		state: {
+			/**
+			 * Get the pagination text.
+			 *
+			 * @return {string} The pagination text showing current page and total pages.
+			 */
+			get paginationText() {
+				const { page, pages } = getContext();
+				return `${ page } / ${ pages }`;
+			},
+
+			/**
+			 * Check if the previous link should be disabled.
+			 *
+			 * @return {boolean} True if the previous link should be disabled.
+			 */
+			get disablePreviousLink() {
+				const { page } = getContext();
+				return page <= 1;
+			},
+
+			/**
+			 * Check if the next link should be disabled.
+			 *
+			 * @return {boolean} True if the next link should be disabled.
+			 */
+			get disableNextLink() {
+				const { page, pages } = getContext();
+				return page >= pages;
+			},
+		},
+		actions: {
+			/**
+			 * Fetch actors for the current page.
+			 *
+			 * @return {Promise<void>} Promise that resolves when actors are fetched.
+			 */
+			async fetchItems() {
+				const context = getContext();
+				const { userId, page, perPage, order, endpoint } = context;
+
+				// Set loading state.
+				context.isLoading = true;
+
+				try {
+					// Build the API path and parameters.
+					const { namespace } = getConfig();
+					const path = url.addQueryArgs( `/${ namespace }/actors/${ userId }/${ endpoint }`, {
+						context: 'full',
+						per_page: perPage,
+						order,
+						page,
+					} );
+
+					// Use apiFetch to get the data.
+					const { orderedItems, totalItems } = await apiFetch( { path } );
+
+					// Update the context with the new items.
+					context.items = orderedItems.map( ( actor ) => ( {
+						handle: '@' + actor.preferredUsername,
+						icon: actor.icon,
+						name: actor.name || actor.preferredUsername,
+						url: validateUrl( actor.url || actor.id ),
+					} ) );
+
+					context.total = totalItems;
+					context.pages = Math.ceil( totalItems / perPage );
+				} catch ( error ) {
+					// eslint-disable-next-line no-console -- Log error for debugging.
+					console.error( `Error fetching ${ endpoint }:`, error );
+				} finally {
+					// Clear loading state.
+					context.isLoading = false;
+				}
+			},
+
+			/**
+			 * Navigate to the previous page.
+			 *
+			 * @param {Event} event The click event.
+			 */
+			previousPage( event ) {
+				event.preventDefault();
+				const context = getContext();
+
+				if ( context.page > 1 ) {
+					context.page--;
+					actions.fetchItems();
+				}
+			},
+
+			/**
+			 * Navigate to the next page.
+			 *
+			 * @param {Event} event The click event.
+			 */
+			nextPage( event ) {
+				event.preventDefault();
+				const context = getContext();
+
+				if ( context.page < context.pages ) {
+					context.page++;
+					actions.fetchItems();
+				}
+			},
+		},
+		callbacks: {
+			/**
+			 * Sets the default avatar when the avatar image fails to load.
+			 *
+			 * @param {Object} event The error event.
+			 */
+			setDefaultAvatar( event ) {
+				event.target.src = getConfig().defaultAvatarUrl;
+			},
+		},
+	} );
+
+	return { actions };
+}
